+++ R/basicPieceGrob.R
@@ -124,13 +124,13 @@ piece_filename_helper <- function(directory, piece_side, format, suit, rank, cfg
 #' @rdname basicPieceGrobs
 #' @export
 pyramidTopGrob <- function(piece_side, suit, rank, cfg=pp_cfg()) {
-    g1 <- pieceGrob("pyramid_face",  suit, rank, cfg, use_pictureGrob=TRUE,
+    g1 <- pieceGrob("pyramid_face",  suit, rank, cfg, type="picture",
                     y=0.75, width=1.0, height=0.5, angle=180, name="face")
-    g2 <- pieceGrob("pyramid_back",  suit, rank, cfg, use_pictureGrob=TRUE,
+    g2 <- pieceGrob("pyramid_back",  suit, rank, cfg, type="picture",
                     y=0.25, width=1.0, height=0.5, angle=  0, name="back")
-    g3 <- pieceGrob("pyramid_left",  suit, rank, cfg, use_pictureGrob=TRUE,
+    g3 <- pieceGrob("pyramid_left",  suit, rank, cfg, type="picture",
                     x=0.25, width=1.0, height=0.5, angle=-90, name="left")
-    g4 <- pieceGrob("pyramid_right", suit, rank, cfg, use_pictureGrob=TRUE,
+    g4 <- pieceGrob("pyramid_right", suit, rank, cfg, type="picture",
                     x=0.75, width=1.0, height=0.5, angle= 90, name="right")
     grobTree(g3, g4, g1, g2, cl="pyramid_top")
 }